backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
spacy==3.7.2
python-docx==1.1.0
PyPDF2==3.0.1
pdfplumber==0.10.3
google-generativeai==0.3.0
python-multipart==0.0.6
requests==2.31.0


pathlib>=1.0.1

